requirements.txt
flask
langchain
langchain-core
openai
python-dotenv
werkzeug
flask_cors
PyMuPDF
PyPDF2
requests
tqdm
gunicorn
langchain-community
langchain-text-splitters
tiktoken
faiss-cpu
beautifulsoup4
langchain-openai
